Move encoding functions to utils
